_soname_deps_qa: Add new check for missing (certain) dependencies from NEEDED
This adds a new QA check which checks the NEEDED entries for to-be-installed
ELF files.

It accepts a list of packages with their SONAMEs and the corresponding
dependency (and whether it should be depended on with a subslot operator).

Future work may extend this to check _all_ packages (as in: check all
ELF files for NEEDED entries where the package providing each element
is not explicitly depended on).

The reason for the more primitive version now is:
- we're looking at performing the libxcrypt migration soon
- the full check is going to be expensive

Signed-off-by: Sam James <sam@gentoo.org>