lib/_emerge/depgraph: show circular dep cycles _after_ conflicts
We can't solve circular deps automatically and a common rule of
thumb we tell users in support channels is "most fatal errors
should be fixed first, down the bottom". This holds for
e.g. autounmask changes and such but the one exception
is circular deps which are at the top of the output (after
the package list) but _after_ conflicts.

Often, conflicts are just as a result of something else
jamming up dependency resolution, in this case, circular
dependencies.

Let's list circular dependencies closer towards the
bottom of the output to make it more clear it's
not noise & needs action to resolve.

Signed-off-by: Sam James <sam@gentoo.org>