bin/phase-functions.sh: Rework the empty directory culling in __dyn_clean()
In __dyn_clean(), there is a routine in which an attempt is made to
recursively delete empty directories beneath PORTAGE_BUILDDIR, with the
exception of WORKDIR and its descendants. It is shortly followed by a
routine in which an attempt is made to remove PORTAGE_BUILDDIR itself,
just before the function returns.

These routines have several issues which are addressed herewith.

For the first routine, use -files0-from to convey PORTAGE_BUILDDIR.
Doing so circumvents the issue whereby find(1) may treat the first
operand as anything other than a starting point. It should be noted that
this requires findutils >=4.9 and that portage should depend upon it
accordingly.

For the first routine, have find(1) employ depth-first recursion. This
should constitute a stronger guarantee that the directories are
traversed in the intended order, and be less expensive than the reverse
sorting of all pathnames encountered.

For the first routine, have find(1) skip non-empty directories by
specifying the -empty option. There was never any sense in conveying
those to rmdir(1) because to do so could only raise ENOTEMPTY. In turn,
there is no longer any justification for pretending that errors don't
matter. As such, refrain from muting STDERR for the invocation of
rmdir(1).

For the first routine, use -print0 to ensure that the pathnames are
conveyed verbatim. Gone is the munging of LF to NUL.

For the first routine, don't inject the literal value of WORKDIR into
the -regex option-argument. Instead, use bash to filter out the paths
pertaining to WORKDIR in such a way that no pattern injection is
involved.

For both routines, use -- to signify end of options to rmdir(1).

For the second routine, have find(1) assist rmdir(1) in the course of
removing PORTAGE_BUILDDIR. The problem with the prior approach was that
the directory might already have been removed, in which case rmdir(1)
would raise an error that was of no concern. Instead, invoke find(1) in
such a way that PORTAGE_BUILDDIR is matched in the case that is is
empty, then have xargs(1) convey the matched pathname - if any - to
rmdir(1). It should be noted that the reason for favouring xargs(1) over
find -exec is to ensure that only the exit status value of rmdir(1) is
propagated. In turn, and as with the first routine, there is no longer
any justification for pretending that errors don't matter. As such,
refrain from muting STDERR for the invocation of rmdir(1) and dispense
with the arbitrary invocation of the true builtin afterwards.

Note that GNU coreutils provides --ignore-fail-on-non-empty as an
extension to rmdir(1). As far as I'm aware, portage cannot assume the
presence of coreutils. Thus, the avoidance of its use is deliberate.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>