emerge: warn on installed packages in @world without ebuilds available
This is an outrageously common issue for users: they emerge something
a while ago, it gets treecleaned (possibly without the user ever seeing
the p.mask entry if they don't use the particular system often), but Portage
diligently works to keep the package installed and its dependencies satisfied.

Portage is completely in the right here, but it's very easy for users
to not be aware that this is still a constraint they've asked for that is
being honoured.

This is particularly bad if that package has a slot-operator dependency (:=)
on something which has just changed subslot, as it can't be rebuilt (because
no ebuild), and users get confusing conflicts as a result.

Warn loudly, as we do for other broken things in the world file, if an entry
in there has no ebuilds available.

Bug: https://bugs.gentoo.org/911180
Signed-off-by: Sam James <sam@gentoo.org>